Add TypeScript
