--- components/Social/SocialCard.module.css
@@ -1,31 +1,0 @@
-.card_container {
-  /* display: flex; */
-  height: 3rem;
-  max-width: 14rem;
-  margin: 0.5rem;
-}
-
-.social_link {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  border: 2px solid;
-  border-radius: 4px;
-}
-
-.icon_container {
-  display: flex;
-  height: 60%;
-  justify-content: center;
-  align-items: center;
-  margin: 1rem;
-}
-
-.social_icon {
-  width: 100%;
-  height: 100%;
-  /* height: 60%; */
-  font-size: 2rem;
-}
