+++ package.json
@@ -24,7 +24,9 @@
     "react-icons": "^4.3.1"
   },
   "devDependencies": {
+    "@types/node": "^18.0.6",
     "eslint": "8.8.0",
-    "eslint-config-next": "12.0.10"
+    "eslint-config-next": "12.0.10",
+    "typescript": "^4.7.4"
   }
 }
