--- pages/blog/[slug].js
@@ -1,47 +1,0 @@
-import fs from "fs";
-import { serialize } from "next-mdx-remote/serialize";
-import { MDXRemote } from "next-mdx-remote";
-
-import path from "path";
-
-import matter from "gray-matter";
-
-export const getStaticPaths = async () => {
-  const files = fs.readdirSync(path.join("posts"));
-  const paths = files.map((filename) => ({
-    params: {
-      slug: filename.replace(".mdx", ""),
-    },
-  }));
-  return {
-    paths,
-    fallback: false,
-  };
-};
-
-export const getStaticProps = async ({ params: { slug } }) => {
-  const markdownWithMeta = fs.readFileSync(
-    path.join("posts", slug + ".mdx"),
-    "utf-8"
-  );
-  const { data: frontMatter, content } = matter(markdownWithMeta);
-  const mdxSource = await serialize(content);
-  return {
-    props: {
-      frontMatter,
-      slug,
-      mdxSource,
-    },
-  };
-};
-
-const PostPage = ({ frontMatter: { title }, mdxSource }) => {
-  return (
-    <div className="mt-4">
-      <h1>{title}</h1>
-      <MDXRemote {...mdxSource} />
-    </div>
-  );
-};
-
-export default PostPage;
